+++ src/style.css
@@ -65,6 +65,7 @@ input{
 /* MAIN HOME */
 .home{
     width: 100%;
+    /*height: 71vh; aplicar media query*/
     display: flex;
     flex-direction: row;
     align-content: flex-start;
@@ -495,18 +496,18 @@ align-items: center;
 }
 
 /* IMAGEN DE EVOLUCIONES */
-.imagenEvolucion{
+.imagenEvolucion1, .imagenEvolucion2{
     width: 80px;
 }
 
 /* CONTENEDOR DEL NOMBRE DE LA PRIMERA EVOLUCION  */
-.nombreEvolucion1{
+.nombreEvolucion1, .nombreEvolucion2{
     display: flex;
     flex-direction: row;
 }
 
 /* NOMBRE DE LA EVOLUCION */
-.nomEvolucion{
+.nomEvolucion1{
     padding-inline: 12px;
 }
 
@@ -557,7 +558,8 @@ align-items: center;
     width: 75%;
     background-color: rgba(255,255,255,0.4);
     border-radius: 30px;
-    padding: 15px;
     text-align: center;
     border: 4px solid rgb(6, 32, 70);
+    padding: 15px;
+    margin: 30px;
 }
